+++ src/server/src/model/week.js
@@ -1,0 +1,11 @@
+const week = [
+    {id:0,name:"Domingo"},
+    {id:1,name:"Segunda-feira"},
+    {id:2,name:"Terça-feira"},
+    {id:3,name:"Quarta-feira"},
+    {id:4,name:"Quinta-feira"},
+    {id:5,name:"Sexta-feira"},
+    {id:6,name:"Sábado"},
+]
+
+module.exports = week